Extract @ObjectId from @Id which is now just an alias for @JsonProperty("_id")
