+++ src/main/java/org/jongo/marshall/jackson/JacksonIdFieldSelector.java
@@ -17,9 +17,8 @@
 package org.jongo.marshall.jackson;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.bson.types.ObjectId;
 import org.jongo.ReflectiveObjectIdUpdater;
-import org.jongo.marshall.jackson.id.Id;
+import org.jongo.marshall.jackson.oid.Id;
 
 import java.lang.reflect.Field;
