+++ src/main/java/org/jongo/marshall/jackson/oid/Id.java
@@ -14,25 +14,18 @@
  * limitations under the License.
  */
 
-package org.jongo.marshall.jackson.id;
+package org.jongo.marshall.jackson.oid;
 
 import com.fasterxml.jackson.annotation.JacksonAnnotationsInside;
-import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 import java.lang.annotation.Retention;
 
-import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @Retention(RUNTIME)
 @JacksonAnnotationsInside
 
-@JsonInclude(NON_NULL)
 @JsonProperty("_id")
-@JsonSerialize(using = ObjectIdSerializer.class)
-@JsonDeserialize(using = ObjectIdDeserializer.class)
 public @interface Id {
 }
