+++ src/main/java/org/jongo/marshall/jackson/oid/ObjectIdSerializer.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright (C) 2011 Benoit GUEROUT <bguerout at gmail dot com> and Yves AMSELLEM <amsellem dot yves at gmail dot com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jongo.marshall.jackson.oid;
+
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+import org.bson.types.ObjectId;
+
+import java.io.IOException;
+
+public class ObjectIdSerializer extends JsonSerializer<String> {
+
+    @Override
+    public void serialize(String value, JsonGenerator jgen, SerializerProvider provider) throws IOException {
+        jgen.writeObject(new ObjectId(value));
+    }
+}
